LayoutTests/http/tests/referrer-policy-iframe/strict-origin/cross-origin-http-UpgradeMixedContent.https.html

<!DOCTYPE html><!-- webkit-test-runner [ UpgradeMixedContentEnabled=true ] -->
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of strict-origin referrer policy when cross origin from HTTPS to HTTP.");
jsTestIsAsync = true;

if (window.internals) {
    internals.settings.setAllowDisplayOfInsecureContent(true);
    internals.setTrackingPreventionEnabled(false);
}

window.onmessage = function(event) {
    testFailed("Document should not be loaded");
    referrer = event.data.referrer;
    // Should be the empty string because we are going from HTTPS to HTTP.
    shouldBeEqualToString("referrer", "");
    if (window.internals)
        internals.setTrackingPreventionEnabled(true);
    finishJSTest();
}
setTimeout(() => {
    testPassed("Loading insecure iframe timed out");
    finishJSTest();
}, 500);
</script>
<iframe src="http://localhost:8000/referrer-policy/resources/document.html" referrerpolicy="strict-origin"></iframe>
</body>
</html>
